various changes
Update Workspace.tscn:
This changes the naming of node stuff associated with music playback in the scene tree to reflect the addition of MusicGroup.

Update MusicGroup:
Since the Song instances are automatically created by MusicGroup's constructor, it makes sense to make MusicGroup's Dispose function dispose of the Song instances created as well.