Update Interactive.cs
This removes the default start and stop code where the clock would be started or stopped.

The idea behind this is that multiple interactives should be able to sync themselves with one clock. In this case, it would not make sense to have multiple interactives start the same clock multiple times. The same concept applies for stopping the clock.