change WIP section of the README
The first version of the game's climbing implementation has been finished, so we need to change the README to account for that.

Climbing logic is still likely to receive changes though. With the current climbing logic, the range of angles you have to face a ladder at is a little too strict. Perhaps we should make the Climber class use two raycasts instead of one.