add some moveable objects
Moveable objects in this commit means platforms that can be manually moved by having the player's character push the object.

In the lobby, moveable objects are  currently indicated with the color yellow.

This is the commit that adds the test pushbox mentioned in the notes (I decided to commit the changes to the notes before making this commit).