replace ProjectSettings.GetSetting usages
The ProjectSettings.GetSetting method only returns "compiled" values that the game boots up with. It doesn't take into account values that could be changed at "runtime".

Therefore, the variables for physics refresh rate and gravity have to be accessed in a way that accounts for changes to these variables at runtime.