Change naming of BaseMover's BodyState enums
The use of all caps for enum names is incorrect according to the .NET style guide.

Yes, this renaming is a breaking change. But I don't think anyone is currently using Jumpvalley's code as a codebase for their own projects right now.

This kind of renaming should be done sooner rather than later so if others decide to use Jumpvalley's code as a codebase, this change does not break their code (assuming the version of Jumpvalley they base their code on is later than v0.3.1).