Update AnimatedNodeGroup.cs
This commit makes it so if the developer wants a limit to the number of nodes currently visible at a time in the group, there can only be one visible at a time.

This was done for thread-safety reasons. This also allows for ease of use since it'd be convenient if you could directly toggle one of the node's IsVisible property and AnimatedGroupNode would still respond to that.