remove BottomRayCollider stuff
The BottomRayColliders as SeparationRayShape3Ds were causing problems. Specifically, the character was able to climb a wall in cases where it shouldn't. Keeping these BottomRayColliders as SeparationRayShape3Ds, the only way to prevent the abnormal wall climbing was to make the BottomRayColliders occupy a smaller area. However, this made long jumping feel more awkward, as the player expects to be able to use the full width and length of the character to be able to perform a long jump. Therefore, the best way to move forward would be to implement Celeste-style pixel corrrection a different way.