change how force application position is calculated
Numerous changes have been made in this commit to the logic used to allow the character to push rigid bodies. This was done to make the force application position the average of the collision positions, as this makes for a better guess of the force application position that the user desires.