allow different collision hitbox shapes for... ...climbing
Before this commit, climbing would only work properly if the shape of the collision hitbox was a rectangular prism. After this commit, the collision hitbox shape can be a capsule (or some other shape where the Axis-Aligned Bounding Box of the character can be calculated) and climbing would still work properly.